moved fetch logic, create spotlight page ad split spotlight function
