+++ pages/index.js
@@ -1,34 +1,10 @@
-import useSWR from "swr";
-import ArtPieces from "@/components/ArtPieces";
 import Spotlight from "@/components/Spotlight";
 
-const URL = "https://example-apis.vercel.app/api/art";
-
-const fetcher = async (url) => {
-  const response = await fetch(url);
-  if (!response.ok) {
-    const error = new Error("Failed to fetch data");
-    error.info = await response.json();
-    error.status = response.status;
-    throw error;
-  }
-  return response.json();
-};
-
-export default function HomePage() {
-  const { data, error, isLoading } = useSWR(URL, fetcher);
-
-  if (error) {
-    return <div>failed to load</div>;
-  }
-  if (isLoading) {
-    return <div>Loading...</div>;
-  }
+export default function SpotlightPage({ pieces }) {
   return (
     <div>
       <h1>Art Gallery App</h1>
-      <ArtPieces pieces={data} />
-      <Spotlight pieces={data} />
+      <Spotlight pieces={pieces} />
     </div>
   );
 }
